Refactor folder stack handling
