+++ docking/ui/folder/__init__.py
@@ -1,0 +1,1 @@
+"""Folder stack UI helpers."""
